fixed migrations

+++ database/migrations/2022_08_03_093413_create_company_addresses_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('company_addresses', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('company_id');
+            $table->unsignedBigInteger('company_id');
             $table->string('name');
             $table->enum('type', ['fiscal', 'logistic', 'administrative']);
             $table->string('street');
